Surpress pointer to uint32_t cast warning.
Since these code is only for ARM architecture, we should use
32bit unsigned integer to store address.  However, this casting
may cause compilation warning on the other platform.  So we are
going to cast to uintptr_t then truncate to uint32_t.

+++ include/impl/ELFObject.hxx
@@ -202,9 +202,9 @@ relocateARM(void *(*find_sym)(void *context, char const *name),
         }
 
         //LOGI("Function %s: using stub %p\n", sym->getName(), stub);
-        S = (uint32_t)stub;
+        S = (uint32_t)(uintptr_t)stub;
 #else
-        S = (uint32_t)callee_addr;
+        S = (uint32_t)(uintptr_t)callee_addr;
 #endif
 
         // Relocate the R_ARM_CALL relocation type
